README
This file provides sequential importance resampling particle filters for two specific gene reaction networks. The filter is established based on hybrid approximations of those models and can accurately approximate the actual filter when the particle population is large. A paper on  the derivation and analysis of the filter will come later.